feat(integrator_filter): extended average to act as integrator
